requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
flask-mail==0.10.0
qrcode[pil]==8.2
Pillow==12.0.0

